requirements.txt
# --- API & Web Framework ---
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
python-multipart==0.0.21
requests==2.32.5
python-dotenv==1.2.1

# --- LangChain & LangGraph Ecosystem ---
langchain==1.2.7
langchain-core==1.2.7
langchain-community==0.4.1
langchain-classic==1.0.1
langchain-openai==1.1.7
langchain-cohere==0.5.0
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.0
langgraph==1.0.7

# --- Vector DBs & Search ---
qdrant-client==1.16.2
fastembed==0.3.3
cohere==5.20.2
rank-bm25==0.2.2

# --- Document Processing (From previous RAG) ---
pypdf==6.6.0
docx2txt==0.9
tiktoken==0.12.0

# --- NEW: Data Analyst Agent Requirements ---
duckdb>=1.0.0
faiss-cpu>=1.8.0
pandas>=2.2.0
openpyxl>=3.1.2